+++ src/cli/args.rs
@@ -6,6 +6,7 @@ use clap::{Parser, ValueEnum};
 pub enum Solver {
     Z3,
     CVC5,
+    Bitwuzla,
     NoSmt,
 }